fix: address CodeQL alerts for workflow permissions and integer overflow
- Added permissions: contents: read to ci.yml and gpu-tests.yml
- Fixed integer overflow in EmbeddingExtractor.h totalSize()
- Fixed integer overflow in StreamingDataLoader.cpp batch size calculation
- Fixed integer overflow in TestKNearestNeighbors.cpp vector allocation

+++ src/amazon/dsstne/pipeline/EmbeddingExtractor.h
@@ -50,7 +50,7 @@ struct EmbeddingResult {
     
     EmbeddingResult() : gpuEmbeddings(nullptr), batchSize(0), embeddingDim(0) {}
     
-    size_t totalSize() const { return batchSize * embeddingDim; }
+    size_t totalSize() const { return static_cast<size_t>(batchSize) * static_cast<size_t>(embeddingDim); }
 };
 
 /**
